+++ sh/common_secrets.sh
@@ -8,13 +8,17 @@ if [ ! -f "$project_root"/secrets.json ] ; then
     exit 1
 fi
 
-if [[ $(ls -l secrets.json | cut -d' ' -f1 | cut -d. -f1) != '-rw-------' ]] ; then
+declare secrets_permissions
+secrets_permissions="$(ls -l "$project_root"/secrets.json)"
+secrets_permissions="${secrets_permissions::10}"
+declare -r secrets_permissions
+if [[ $secrets_permissions != '-rw-------' ]] ; then
     echo 'secrets.json permissions too lax' >&2
     echo 'run: chmod 600 secrets.json' >&2
     exit 1
 fi
 
-if ! sha256sum -c <<<'bb82de121880f1182dbae410b341749e5ac1355954ae6c03151a1826e7bba745  secrets.json' >/dev/null ; then
+if ! sha256sum -c <<<'8c340c2ab35e244f8535f27f94f939a868e9416a013496cc61d9874342b697c6  secrets.json' >/dev/null ; then
     echo 'Secrets are wrong' >&2
     exit 1
 fi