+++ src/core/UniswapV4.sol
@@ -11,9 +11,7 @@ import {UnsafeMath} from "../utils/UnsafeMath.sol";
 
 import {TooMuchSlippage, DeltaNotPositive, DeltaNotNegative, ZeroSellAmount} from "./SettlerErrors.sol";
 
-import {
-    BalanceDelta, IHooks, IPoolManager, UnsafePoolManager, IUnlockCallback
-} from "./UniswapV4Types.sol";
+import {BalanceDelta, IHooks, IPoolManager, UnsafePoolManager, IUnlockCallback} from "./UniswapV4Types.sol";
 import {Encoder, NotesLib, StateLib, Decoder, Take} from "./FlashAccountingCommon.sol";
 
 library CreditDebt {